[V1] Fix int32 overflow in grammar-bitmask structured-outputs kernel
Promote logits_idx to int64 at load before the `logits_idx * logits_stride`
store offset so spec-decode-expanded logits rows are addressed correctly past
2**31 elements.

+++ vllm/v1/worker/gpu/structured_outputs.py
@@ -93,7 +93,7 @@ def _apply_grammar_bitmask_kernel(
     BLOCK_SIZE: tl.constexpr,
 ):
     bitmask_idx = tl.program_id(0)
-    logits_idx = tl.load(logits_indices_ptr + bitmask_idx)
+    logits_idx = tl.load(logits_indices_ptr + bitmask_idx).to(tl.int64)
 
     # Load the bitmask.
     block_id = tl.program_id(1)
